fix: change TotalAmount type from float to int in TransactionDTOs for consistency

+++ DTOs/TransactionDTOs.cs
@@ -9,7 +9,7 @@ public class TransactionDTOs
     public Guid ProductId { get; set; }
     public required string ProductName { get; set; }
     public Guid? UserId { get; set; }
-    public float TotalAmount { get; set; }
+    public int TotalAmount { get; set; }
     public int Quantity { get; set; }
     public int Price { get; set; }
     }
